le générateur fonctionne sans segfault ; les str sotn encore null ; les nodes pipe ne s'ajoutent pas correctement ; certainement des problèmes de free, pas importants pour le moment
