+++ .vscode/settings.json
@@ -1,3 +1,7 @@
 {
-	"C_Cpp.errorSquiggles": "disabled"
+	"C_Cpp.errorSquiggles": "disabled",
+	"files.associations": {
+		"all_struct.h": "c",
+		"fn_dev.h": "c"
+	}
 }
